New Bug1 Mode; uses "Key Down" event to manage/set "decodeval" value
This update debuts a new stategy in CW decoding. Its intended to better
handle keying timing often found with "Bug" style sending. It differs
from previous algorithms that have relied on timing linked to "key up"
event. It instead uses the "Key down" to decide if this i a continuation
of the charater being sent, or truely a new charater. This stategy is
enabled in the "Bug1" mode
At this time, its still, "a work in progress".